gurt

+++ CUE4Parse/UE4/Objects/UObject/UState.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using CUE4Parse.UE4.Assets.Readers;
 using CUE4Parse.UE4.Versions;
-using CUE4Parse.Utils;
 using Newtonsoft.Json;
 
 namespace CUE4Parse.UE4.Objects.UObject;
@@ -34,7 +33,7 @@ public class UState : UStruct
         }
         else
         {
-            ProbeMask = Ar.Read<long>();
+            ProbeMask = Ar.Read<int>();
         }
 
         LabelTableOffset = Ar.Read<short>();
